Consolidate files
`npm i` fails locally, need to investigate
